fixing bugs connected long type for factorial and one loops for AdditionalTask1

+++ src/main/java/net/vrakin/BasicTasks.java
@@ -2,8 +2,8 @@ package net.vrakin;
 
 public class BasicTasks {
 
-    public static int factorial(int n) {
-        int result = 1;
+    public static long factorial(int n) {
+        long result = 1;
         for (int i = n; i >= 1; i--) {
             result *= i;
         }
